requirements.txt
pydantic~=2.4.2
SQLAlchemy~=2.0.22
fastapi~=0.104.0
passlib~=1.7.4
python-dotenv~=1.0.0
PyYAML~=6.0.1
uvicorn~=0.23.2
pydantic[email]~=2.4.2
python-decouple==3.8
pydantic_settings==2.0.3
psycopg2-binary==2.9.9
alembic==1.12.0
starlette==0.27.0
python-multipart==0.0.6
python-jose==3.3.0
